--- a/attribute.go
+++ b/attribute.go
@@ -4,27 +4,33 @@
 package rpsl
 
 import (
+	"errors"
 	"fmt"
 	"strings"
 )
 
+// Attribute represents a parsed attribute with a normalized key.
 type Attribute struct {
 	Name  string
 	Value string
 }
 
+// newAttribute creates an Attribute from a name and value string,
+// normalizing the key to lowercase and cleaning the value.
 func newAttribute(name, value string) Attribute {
 	key := strings.ToLower(name)
-
-	var cleanedLines []string
+	cleanedLines := make([]string, 0, 4)
 	lines := strings.Split(value, "\n")
+
 	for i, line := range lines {
+		// Remove a leading '+' from continuation lines.
 		if i > 0 && strings.HasPrefix(line, "+") {
 			line = line[1:]
 		}
 
-		if strings.Contains(line, "#") {
-			line = strings.Split(line, "#")[0]
+		// Remove inline comments by splitting on '#' and taking the first part.
+		if idx := strings.IndexByte(line, '#'); idx >= 0 {
+			line = line[:idx]
 		}
 
 		line = strings.TrimSpace(line)
@@ -33,13 +39,16 @@ func newAttribute(name, value string) Attribute {
 		}
 	}
 
-	value = strings.Join(cleanedLines, " ")
-	return Attribute{Name: key, Value: value}
+	// Join the cleaned lines with a space.
+	cleanedValue := strings.Join(cleanedLines, " ")
+
+	return Attribute{Name: key, Value: cleanedValue}
 }
 
+// parseAttributes parses the given buffer into a slice of Attributes.
 func parseAttributes(buf string) ([]Attribute, error) {
 	if buf == "" {
-		return nil, fmt.Errorf("object cannot be null")
+		return nil, errors.New("parseAttributes: object cannot be null")
 	}
 
 	var attributes []Attribute
@@ -47,11 +56,11 @@ func parseAttributes(buf string) ([]Attribute, error) {
 
 	for pos < len(buf) {
 		key, newPos, err := parseKey(buf, pos)
-		pos = newPos
 		if err != nil {
-			return nil, err
+			return nil, fmt.Errorf("parseAttributes: %w", err)
 		}
 
+		pos = newPos
 		value, newPos := parseValue(buf, pos)
 		pos = newPos
 
@@ -61,36 +70,44 @@ func parseAttributes(buf string) ([]Attribute, error) {
 	return attributes, nil
 }
 
+// isValidKeyChar returns true if c is allowed in a key.
 func isValidKeyChar(c byte) bool {
-	return (c >= 'a' && c <= 'z') || (c >= 'A' && c <= 'Z') || (c >= '0' && c <= '9') || c == '-' || c == '*'
+	return (c >= 'a' && c <= 'z') ||
+		(c >= 'A' && c <= 'Z') ||
+		(c >= '0' && c <= '9') ||
+		c == '-' || c == '*'
 }
 
+// isLineContinuationChar returns true if c signifies a continuation of a line.
 func isLineContinuationChar(c byte) bool {
 	return c == ' ' || c == '\t' || c == '+'
 }
 
+// parseKey extracts a key ending at the first ':' and returns the key,
+// the position after the colon, and an error if any.
 func parseKey(buf string, pos int) (string, int, error) {
 	start := pos
+
 	for pos < len(buf) {
 		c := buf[pos]
 		if c == ':' {
 			if pos == start {
-				return "", 0, fmt.Errorf("read zero sized key")
+				return "", 0, fmt.Errorf("parseKey: zero-sized key at pos %d", pos)
 			}
-
 			return buf[start:pos], pos + 1, nil
 		}
 
 		if !isValidKeyChar(c) {
-			return "", 0, fmt.Errorf("read illegal character in key: '%c'", c)
+			return "", 0, fmt.Errorf("parseKey: illegal character '%c' at pos %d", c, pos)
 		}
 
 		pos++
 	}
 
-	return "", 0, fmt.Errorf("no key found")
+	return "", 0, fmt.Errorf("parseKey: no key found starting at pos %d", start)
 }
 
+// parseValue extracts a value until a newline that is not followed by a continuation char.
 func parseValue(buf string, pos int) (string, int) {
 	start := pos
 	stop := pos
@@ -108,6 +125,7 @@ func parseValue(buf string, pos int) (string, int) {
 			if isLineContinuationChar(next) {
 				continue
 			}
+
 			break
 		}
 
@@ -117,6 +135,7 @@ func parseValue(buf string, pos int) (string, int) {
 	return buf[start:stop], pos
 }
 
+// String returns a string representation of the Attribute.
 func (a *Attribute) String() string {
 	var str strings.Builder
 	str.WriteString(a.Name)