--- a/object.go
+++ b/object.go
@@ -4,6 +4,7 @@
 package rpsl
 
 import (
+	"errors"
 	"fmt"
 	"strings"
 )
@@ -12,7 +13,7 @@ type Object struct {
 	Attributes []Attribute
 }
 
-// Returns a slice of unique keys present in the Object.
+// Keys returns a slice of unique keys present in the Object.
 // If a key appears multiple times in the Object, it will only be included once in the returned slice.
 func (o *Object) Keys() []string {
 	keyPresent := make(map[string]struct{})
@@ -27,12 +28,12 @@ func (o *Object) Keys() []string {
 	return keyList
 }
 
-// Returns the number of attributes in the Object.
+// Len returns the number of attributes in the Object.
 func (o *Object) Len() int {
 	return len(o.Attributes)
 }
 
-// Returns the first value for a given key in the Object.
+// GetFirst returns the first value for a given key in the Object.
 // If the key is not present in the Object, an empty string will be returned.
 // If a key appears multiple times in the Object, only the first value will be returned.
 func (o *Object) GetFirst(key string) *string {
@@ -46,7 +47,7 @@ func (o *Object) GetFirst(key string) *string {
 	return nil
 }
 
-// Returns a slice of values for a given key in the Object.
+// GetAll returns a slice of values for a given key in the Object.
 // If the key is not present in the Object, an empty slice will be returned.
 // If a key appears multiple times in the Object, all values will be included in the returned slice.
 func (o *Object) GetAll(key string) []string {
@@ -60,7 +61,7 @@ func (o *Object) GetAll(key string) []string {
 	return attributes
 }
 
-// Returns true if the Object contains a given key.
+// Exists returns true if the Object contains a given key.
 func (o *Object) Exists(key string) bool {
 	key = strings.ToLower(key)
 	for _, attr := range o.Attributes {
@@ -72,7 +73,7 @@ func (o *Object) Exists(key string) bool {
 	return false
 }
 
-// Returns a string representation of the Object.
+// String returns a string representation of the Object.
 func (o *Object) String() string {
 	var str strings.Builder
 
@@ -85,10 +86,10 @@ func (o *Object) String() string {
 	return str.String()
 }
 
-// Ensures that the first attribute in the Object is of a given class.
+// EnsureClass ensures that the first attribute in the Object is of a given class.
 func (o *Object) EnsureClass(class string) error {
 	if len(o.Attributes) == 0 {
-		return fmt.Errorf("object has no attributes")
+		return errors.New("object has no attributes")
 	}
 
 	first := o.Attributes[0].Name
@@ -99,7 +100,7 @@ func (o *Object) EnsureClass(class string) error {
 	return nil
 }
 
-// Ensures that the Object has at least one attribute with a given key.
+// EnsureAtLeastOne ensures that the Object has at least one attribute with a given key.
 func (o *Object) EnsureAtLeastOne(key string) error {
 	if !o.Exists(key) {
 		return fmt.Errorf("attribute '%s' is (mandatory, multiple) but found none", key)
@@ -108,7 +109,7 @@ func (o *Object) EnsureAtLeastOne(key string) error {
 	return nil
 }
 
-// Ensures that the Object has at most one attribute with a given key.
+// EnsureAtMostOne ensures that the Object has at most one attribute with a given key.
 func (o *Object) EnsureAtMostOne(key string) error {
 	if len(o.GetAll(key)) > 1 {
 		return fmt.Errorf("attribute '%s' is (optional, single) but found multiple", key)
@@ -117,7 +118,7 @@ func (o *Object) EnsureAtMostOne(key string) error {
 	return nil
 }
 
-// Ensures that the Object has exactly one attribute with a given key.
+// EnsureOne ensures that the Object has exactly one attribute with a given key.
 func (o *Object) EnsureOne(key string) error {
 	if err := o.EnsureAtLeastOne(key); err != nil {
 		return fmt.Errorf("attribute '%s' is (mandatory, single) but found none", key)
@@ -131,34 +132,42 @@ func (o *Object) EnsureOne(key string) error {
 }
 
 func parseObjects(buf string) ([]Object, error) {
-	objects := []Object{}
+	var objects []Object
 
 	if buf == "" {
 		return objects, nil
 	}
 
 	lines := strings.Split(buf, "\n")
-	for i, line := range lines {
-		if strings.HasPrefix(line, "%") || strings.HasPrefix(line, "#") {
-			lines[i] = ""
-		}
-	}
 
-	buf = strings.Join(lines, "\n")
-	for _, part := range strings.Split(buf, "\n\n") {
-		part = strings.TrimPrefix(part, "\n")
-		part = strings.TrimSuffix(part, "\n")
-		if part == "" {
+	// Process line by line, accumulating non-comment lines.
+	var currentPart []string
+
+	for i := 0; i <= len(lines); i++ {
+		// Process a line or handle end of input.
+		isEndOfFile := i == len(lines)
+		isEmptyLine := !isEndOfFile && lines[i] == ""
+
+		// When we hit an empty line or EOF, process the accumulated part.
+		if isEmptyLine || isEndOfFile {
+			if len(currentPart) > 0 {
+				partText := strings.Join(currentPart, "\n")
+				attributes, err := parseAttributes(partText)
+				if err != nil {
+					return nil, err
+				}
+
+				objects = append(objects, Object{Attributes: attributes})
+				currentPart = currentPart[:0] // Clear slice without reallocating.
+			}
 			continue
 		}
 
-		attributes, err := parseAttributes(part)
-		if err != nil {
-			return nil, err
+		// Skip comment lines.
+		line := lines[i]
+		if !strings.HasPrefix(line, "%") && !strings.HasPrefix(line, "#") {
+			currentPart = append(currentPart, line)
 		}
-
-		object := Object{Attributes: attributes}
-		objects = append(objects, object)
 	}
 
 	return objects, nil