Hey! Firstly, great project, we rely on it for some of our core data processing :)

We noticed an issue with our project where memory wasn't being returned from the heap. It ended up being due to holding strings returned by `ParseMany`, more specifically the `parseObjects` function. We were able to fix this on our end by explicitly nulling data after use, but I thought I'd take some time to dive into your library to see if I could address the issue.

This PR:
- Fixes the heap issue (by slightly reworking the `parseObjects` function
- Reduces the memory allocations from the tests from 35M objects down to 19.5M objects (`newAttribute` was the main culprit)
- Adds a test to cover the `object.GetFirst` function
- Enhances a few internal errors with a little more context
- Updates exposed function comments to fix the [Go docs guidelines](https://tip.golang.org/doc/comment)

Original library memory profile (from tests):
<img width="676" alt="Screenshot 2025-03-19 at 19 19 25" src="https://github.com/user-attachments/assets/68aff8ea-8a5a-4b95-b6b5-c0fe945de79d" />

Revised library memory profile (from tests):
<img width="665" alt="image" src="https://github.com/user-attachments/assets/11fe3674-4c07-4d2a-ae9c-76899e0ebcb0" />

From our testing, this looks good and doesn't cause any regressions. But please do test it, and let me know if anything should be changed!