@wolveix 2025-03-19
I just added another commit which further optimized the attribute function, but more importantly added a `ParseManyFromReader` function. I think this should actually replace the current `ParseMany` function as it's much more efficient; however, I didn't want to change your package's API without first getting your thoughts :) But in our project test, it dropped our allocs from 60GB down to 43GB.