@frederic-arr 2025-03-19
Hi, thanks for the PR! Everything looks good on my end.

> I just added another commit which further optimized the attribute function, but more importantly added a ParseManyFromReader function. I think this should actually replace the current ParseMany function as it's much more efficient; however, I didn't want to change your package's API without first getting your thoughts :) But in our project test, it dropped our allocs from 60GB down to 43GB.

I don't mind introducing that method but `ParseMany` needs to remain for backwards compatibility reasons. I sense a code duplication by keeping the two so replacing the body of `ParseMany` to a call to `ParseManyFromReader` may be necessary. If we follow this approach `parseObjects` should be removed too as it would be otherwise unused.

---

If this sounds good to you, and if you have time, making the required changes would be much appreciated but I can take them on otherwise.