@wolveix 2025-03-20
Back to the drawing board :) I'm not too sure why the stream reader causes allocs to increase so drastically, it should be the other way round