@wolveix 2025-03-20
This took a lot of back and forth to figure out the optimal implementation, but I think I've found it. The benchmark shows that parsing is faster than before:

```
BenchmarkParseFromReader-12    	 1288480	       914.8 ns/op
```

Memory allocs are slightly higher within the library (~19M), but allow the caller to drastically reduce their allocs; in our project, being able to use `io.Reader` drops our initial sync from 60GB allocs down to just 18GB (60GB was from the previous commit).

As tests fluctuate from one run to another, I was able to see the memory profile for the tests from to 16.9M:
![image](https://github.com/user-attachments/assets/8b950966-3803-4ae3-b8f7-048ce99d6a66)

---

Hopefully this looks good, but let me know if you want anything revised!