@wolveix 2025-03-20
Figured I'd revisit the `newAttribute` function as that's where the majority of the heavy allocs occur. I was able to further optimize it:

```
BenchmarkParseFromReader-12    	 1369954	       849.3 ns/op
```

Memory profiles are now averaging 18M again :)