@wolveix 2025-03-21
Okay, I think I've squeezed every bit of performance I can at this point.

```
BenchmarkParseFromReader-12    	 1547264	       760.8 ns/op
```

Memory allocs remain roughly the same as before, but I also made a few micro optimizations:
```
Attribute.String() optimization:
28ns -> 15 ns/op

Object.String() optimization:
104ns -> 75 ns/op
```

Thanks for making such a great library, and hopefully there are no issues with these changes!